Reduced dependencies.
