+++ src/util/mod.rs
@@ -15,29 +15,10 @@
 // specific language governing permissions and limitations
 // under the License.
 
-/// Converts numeric type to a `String`
-#[inline]
-pub fn lexical_to_bytes<N: lexical_core::ToLexical>(n: N) -> Vec<u8> {
-    let mut buf = Vec::<u8>::with_capacity(N::FORMATTED_SIZE_DECIMAL);
-    unsafe {
-        // JUSTIFICATION
-        //  Benefit
-        //      Allows using the faster serializer lexical core and convert to string
-        //  Soundness
-        //      Length of buf is set as written length afterwards. lexical_core
-        //      creates a valid string, so doesn't need to be checked.
-        let slice = std::slice::from_raw_parts_mut(buf.as_mut_ptr(), buf.capacity());
-        let len = lexical_core::write(n, slice).len();
-        buf.set_len(len);
-    }
-    buf
-}
-
-/// Converts numeric type to a `String`
-#[inline]
-pub fn lexical_to_string<N: lexical_core::ToLexical>(n: N) -> String {
-    unsafe { String::from_utf8_unchecked(lexical_to_bytes(n)) }
-}
+#[cfg(any(feature = "compute", feature = "io_csv"))]
+mod lexical;
+#[cfg(any(feature = "compute", feature = "io_csv"))]
+pub use lexical::*;
 
 #[cfg(feature = "benchmarks")]
 pub mod bench_util;
